Fix keyboard navigation in prefs
The browser_style stylesheet hides checkbox and radio button input
elements and recreates the widget as part of the label; this makes them
inaccessible by keyboard.  The browser skin in-content stylesheet
doesn't have this problem, so we use it instead.  (It also looks more
like the rest of the browser.)